Preview page slightly updated
After finishing main layout for level_info Component, preview page is
the next one to design. The following changes are added:

1) Created dynamic Component for each Serie, and then appending all
these into a special serie container. It remains dynamic and it is only
based on all_level object which provides all necessary data.

2) Prepared initial layout for serie container and container items.
For now, it's simply backgrounds and borders. Plus for each level a
fancy colorful circle is added, which would be useful during next
stages of designing.

3) Slightly changed Preview Page Structure. On top of page is now
Home button and the title. There would be also a github profile link
added, but for now a replacement is simply a short description text.