Create object plans and decoration management
This commit brings out new quality for creating and developing levels.
A new piece of code to memory_game component is added, which allows to
handle dynamic CSS predefined styles importing. On top of that, a new
script is bounded, which is responsible for creating objects and then
appending them to a given plan. This script has to be created separately
for each level.

Overall, it is now possible to easily create and manage designing
'formalities', keeping code cleaner and logically separated.