Update level 5
Another level updated for The Flash serie. The old version used to be
clunky, especially for its design and dull gameplay. This version brings
up some vital changes to background layout. Furthermore, this level
introduces 4-click mechanism, which is brand new in game. It simply
means finding 4 tiles with the same icon - which opens up a lot of
possibilities to make further levels with higher complexity.