Add the database & Prisma to the project
Now the first rework step is taken. Now the database is connected to
the project, which is being integrated with Prisma. Created an initial
schema template, from which two sample levels got filled with some
random data and pushed as a very first records inside the database.
Also, the first query to database (getStaticProps) is now successfully
set up - received expected pieces of data.

In further updates, it is going to be essential to add some app routing,
getting rid of passing random props between main components and
redefining the behaviour for Preview page.