Update game engine to interact with database
Made a game mini remake, which resulted in engine being able to take and
handle all pieces of data from the database (instead of storing the
level information in a separate file). Engine is mostly working as it
used to, but there are couple of bugs and errors to resolve. Currently,
it is possible to play levels, but not doing any post-game actions
(like restarting the level, exiting it, etc.)