Develop Progress records ability to update
This is a vital change to the app. Getting one step further to fully
complete the database integration with the project.

The system is now capable not only of creating a Progress records, but
also to update them when necessary. Whenever player makes any kind of
progress (say, user got level completion percentage record), a
respective Progress record has the completion percentage field
updated with a new (best) value. This mechanism is supposed to work
only if there is a value to update, and updates only those values
which user improved on (for example, player got better highscore).

After some testing, the concept works great and currently there is no
need to make any improvements or fixes in this field.