Replace Game component routing system (part 1)
After dealing with some disadvantages of making dynamic (level ID based)
routes for game levels, a need for change appeared.

Basically, this and next commit are going to be about changing the
dynamic routing for state-based rendering (CSR). The reason for that is
not only it makes more sense to keep the levels being rendered on the
client side, but it also reduces some issues that SSR could have
opened (slower page loading, refreshing level which leads to async
imports not always loading on time = errors).

Here I changed old the way level specific props are obtained
(from getServerSideProps function to a props inheritance from
Preview component to the main Memory_Game component).