Replace Game component routing system (part 2)
The process of changing the routing system is completed by now.
The Game component, which is responsible for handling levels play, is
fully CSR. As a result, this slight update greatly enhanced overall
performance and it will prevent many complications regarding the sole
purpose of using ID based routes for each route separately.