Start applying user authentication to app
Currently the last piece remained to update. That is the authentication
part. App will be playable only for users who registers using Google
Authentication Provider. Thanks to the authentication approach,
players can have a safe accounts ("passwordless"), which can store
their in-game progress permanently. One drawback of such approach is
that it is implied / stated, that each player has to have a Google
Account already set up.

This minor commit is just the starting point. So far, Google Auth is
linked correctly to a project (thanks to NextAuth), but there are still
some major improvements to do, such as: getting session data from a
user account who is playing (and identyfying the account among the
others existing) and also some UI changes (mainly to provide more clear
indication, that a given user is currently logged in).