Complete authentication logic
Google authentication is fully designed. This update involved some
changes that finally make the app authentication part work as intended.
Now, there is a possibility to log in for more than one account per
device. A simple change here is that a log out button is now included in
the Landing page More than that, a progress save process got updated,
so that there is no possibility that a player can get a "progress
downgrade" effect, which resulted in overriding "better progress" with
the worse one. Also changed api routing a bit, so that it can handle 2
ID-based parameters in one single query.

Currently, the main logic part of the app is completed (finally!), and
thanks to that, next future updates will be more about correcting the
app layout, sharpening design for elements and so on. Maybe a quick
remodel for levels will be also done, to "sweep up" a little mess
around the properties that any level can have.