Begin build for User Profile sub-component
To add a subtle indicator of which account is currently logged in, a
profile component had to be made. So, with this in mind I decided to
take a step back to focus on this exact functionality. This one
mini-component will not hold anything complex, but rather some basic
pieces of data about the user and his current progresses. In this commit
I did created a simple UI for User Profile without any functionalities
nor animations added. Future update would cover these missing parts.