Finished User Tab component
Finally, the user account component is finally done. As a final part, I
have added some smooth transitions and animations for UserTab show and
hide actions. Moreover, styling got improved a bit and now working on
applying a similiar version of the component to a Landing Page aswell.
It is going to finalize the proces of creating user specific UI.