Add the player level and exp mechanism
This time Memory Flash received another fun feature, which is levelling
up. When a player plays and wins a level, he gets rewarded for that.
The prize is some EXP, which is based mostly on level difficulty and
on the condition if a level is won for the first time by player.

Currently, there are 40 levels in game (which in future may slighlty
change along with exp floors). This commit is all about the things
related to EXP, so now that everything in this matter is completed,
another updates are going to come !