Redefine level model
Most likely this one is the last change for Level model. Got rid of
many properties, which are now bound to more general property names that
are responsible to store more information on its' own. Because of this
modification, a memory game engine got shifted a bit to handle newly
designed level models correctly.