Landing page rework almost complete
It's been some time since the last commit. This one was really time
consuming, since it was about spotting the weaknesses of old-style
Landing page, and working around them - to make the overall design
much more stable and predictable across devices. The design part is
most likely all done, but a little to work left is for the
compatibility part (preparing layout for various devices). Anyhow,
the current work done is still worth posting it, as it really
redefined the look of the Landing Component to the next level.