Remake Level Info component
The v2 Level Info component, which design was attempted by me several
months ago, was a madness. Data used to be placed quite randomly, with
no real pattern nor scheme. Also styling was looking unorganised.
Because of all these flaws, a solid update was essential. This commit is
the first, breaking part, which pretty much sets up the whole new
layout for refreshed Level Info component. Most of the styling is
already within this commit, the only thing to work on is to apply
dynamic data back again (and maybe to add some extra fancy stylings to
the Component).