Added fade animations between components
Interactivity betweem different components inside the app was
quite forgotten, so now the time has come to make up for this
feature. The transitions are just an opacity mainpulation, but
simplicity is the key here. Transitions are now present in each and
every part of the app, so there are no clunky or flickering layout
changes.

Also, detailed level component got finally changed by adding a
showUp, and hide animations. Many improvements were made regarding
this, so it looks eye-pleasant and fun to watch.