Add account ban mechanism
Flash Memory Game now contains a neat way to ban accounts which
violate the app rules. When such account is about to reach the
Play (Preview) Component, player is being served with a UserBanned view
instead. This component contains some brief message to a user, informing
him that his account got banned and a possible solution to perform in
order to - possibly - retrieve back the access to the game.