Add second level to the game
This commit is starting a new chapter in the app development. From now
on, each subsequent commit will focus on deploying either some
levels-related functionalities or new levels on its' own. This time a
new level got added (which is actually 2nd old-version Memory Game level)
but updated to a huge extent.