Complete level 4 + some bugs solving
'Steps to hell' is now the official release for updated Flash MG.
That was pretty tedious to make, simply because of handling a new
feature, which is secrets (and also introducing level variables to the
game). Also, there were some bugs, revealed accidentally when working
on this level. All of them got luckily solved. Hence a new level can
be freely designed without worries - as for now.