Create level 5
'Lights' are finally completed ! The level on its own does not provide
any new mechanic, but it definitely needed some visual update and
performance testing. In terms of the last mentioned, there were a bit
of things to do in order to maintain a balance between the decoration
count and the seamless flow. Now it's all done, and also playtested
quite well, so the level should not hide any unexpected behaviours or
bugs.