Finished level 6
Added another level to the game. It took not that much time in terms
of decorating. Plus it is just a one stage level, so there were not
really much things to do. The real issue was to apply the right flow
and limitations for the brand new difficulty level - the insane.