Create level 8 (11th in the game by order)
After struggling a bit, 'Camo' is finally completed. It is definitely
one of more good-looking levels out there, since it is kept in camo
style and its the first new version level which makes use of an svg as
a background. The level has been playtested quite a lot, so that there
are no bugs related to the very unique mechanics presented here.