Create level 10.
A brand new level is released ! 'Toxic Factory' is not only first level
outside of old MG, which is added to the game, but also it is the first
extreme level ever ! This one took me much more time that I was
expecting, mostly because of SVG image which anyways is still not
completed. Anyhow gameplay and design parts were the things which got
prepared with a lot of caution. So hopefully it's just a matter of not
completed background - which I am going to fully complete soon.