Create level 11.
A brand new level is completed ! Took some time to complete since I did
not have an actual idea of what mechanics the level should contain.
Also a major part of development process took to create an SVG
background, which was fun to make, but it took decent amount of time to
complete. Other than that, level itself does not contain any weird or
extra stuff, so that the end part was quite quick.