Final preparations for application release (part 1)
The final part has just began! The very last details and modifications
were left to be done as a cherry on top of the project. That means some
extra improvements, as well as solving some known issues. This was quite
exhasutive work, but thankfully many different factors are now solved
starting from this commit.