Final preparations for application release (part 2)
Another part of the final changes is now completed. There has been
made some very impactful changes regarding the Landing Page (which
finally looks fine). Also tweaked the favicon to a brand new one.
More than that, some bugs were solved and that it is all for this
preparation part.