A minor workaround for level 1 - 6 backgrounds
As above mentioned levels do not have their SVG backgrounds, a bg
loading event in the level_info Component was showing some unnecessary
errors to the console. To avoid this, some dummy SVG backgrounds are
prepared. They are being loaded, but on top of them the same old
background-exceptional gradients are applied. So the functionality
remains the same, but the misleasing errors are no longer a problem.